🐛 Fixed git status sometimes being out of line on conflicting boundaries
If gitree was run outside of a git repository but given a path inside
a repository as a target, it would fail to retrieve any files.

This was because the internal Git commands run were always relative to the
CWD. These are now run from the target's position.

On top of that, the wrapper directory can now safely span multiple levels
without being split into separate items.